Extract tests to separate module
